+++ src/routes/(v2)/v2/allow-files/+page.server.ts
@@ -1,5 +1,6 @@
 import { zod } from '$lib/adapters/zod.js';
-import { failAndRemoveFiles, message, superValidate } from '$lib/server/index.js';
+import { removeFiles, message, superValidate } from '$lib/server/index.js';
+import { fail } from '@sveltejs/kit';
 import { schema } from './schema.js';
 
 export const load = async () => {
@@ -15,7 +16,7 @@ export const actions = {
 		const form = await superValidate(formData, zod(schema), { allowFiles: true });
 		console.log('POST', form);
 
-		if (!form.valid) return failAndRemoveFiles(400, { form });
+		if (!form.valid) return fail(400, removeFiles({ form }));
 
 		return message(form, 'Posted OK!');
 	}